fix : 메인페이지 bookmark 오류 수정

+++ src/main/java/com/hotel/service/BookmarkService.java
@@ -7,5 +7,5 @@ public interface BookmarkService {
 	
 	int deleteBookmark(String miId, int siId);
 	
-	List<Long> getBookmarkList(String miId);
+	List<Integer> getBookmarkList(String miId);
 }
